requirements.txt
# attr==0.3.2
# colorama==0.4.6
# ConfigParser==7.1.0
# cryptography==42.0.5
# # docutils==0.18.1
# docutils==0.20
# filelock==3.16.1
# HTMLParser==0.0.2
# ipython==8.25.0
# # ipython==8.12.3
# ipywidgets==7.8.1
# jnius==1.1.0
# keyring==24.3.1
# # numpy==2.1.3
# numpy==1.19.3
# opencv_python==4.5.5.62
# opencv_python_headless==4.10.0.84
# Pillow==7.1.0
# # Pillow==11.0.0
# protobuf==3.20.3
# #pyOpenSSL==24.0.0
# pyOpenSSL==24.2.1
# redis==5.2.0
# scikit_learn==1.4.2
# Sphinx==8.1.3
# streamlit==1.32.0
# tensorflow==2.18.0
# # tensorflow_intel==2.18.0
# thread==2.0.5
# urllib3_secure_extra==0.1.0
# xmlrpclib==1.0.1

streamlit
tensorflow
opencv_python_headless
numpy
scikit-learn
Pillow
matplotlib
pandas
pydeck
streamlit-folium
folium
geopy